Ensure moved items always have controller set.
Due to async loading of tree childred, in some cases when moving an item
its controller was getting deleted after being added, i.e. setController
was being called first on the item in the new tree, before the same item in
the old tree was destroyed and deleteController was called.